Refactor App startup and Get Logs page functional  (initial)
