+++ src/DesktopClient/PresenceLight/Pages/LogsPage.xaml
@@ -15,7 +15,37 @@
 
     <ScrollViewer>
         <Grid Margin="24,0">
-            <lc:Logs x:Name="logs"/>
+           
+        <TabControl x:Name="tabLogs"  >
+            <TabItem Header="Live Logs" >
+                <DataGrid x:Name="dgLiveLogs" ItemsSource="{Binding}" AutoGenerateColumns="False" RowHeight="20">
+                    <DataGrid.Columns>
+                        <DataGridTextColumn Header="Timestamp"  Binding="{Binding Timestamp}"/>
+                        <DataGridTextColumn Header="Level" Binding="{Binding Level}" />
+                        <DataGridTextColumn Header="Message" Binding="{Binding MessageTemplate}" />
+
+                    </DataGrid.Columns>
+                </DataGrid>
+            </TabItem>
+            <TabItem Header="Log Files">
+                <DataGrid x:Name="dgLogFiles" ItemsSource="{Binding}" AutoGenerateColumns="False" RowHeight="20">
+                    <DataGrid.Columns>
+                        <DataGridHyperlinkColumn Header="FileName"  Binding="{Binding Name}" >
+                            <DataGridHyperlinkColumn.ElementStyle>
+                                <Style>
+                                    <EventSetter Event="Hyperlink.Click" Handler="DG_Hyperlink_Click"/>
+                                </Style>
+                            </DataGridHyperlinkColumn.ElementStyle>
+                        </DataGridHyperlinkColumn>
+                        <DataGridTextColumn Header="Size" Binding="{Binding Length}" />
+                        <DataGridTextColumn Header="Creation Time" Binding="{Binding CreationTime}" />
+                        <DataGridTextColumn Header="Last Access" Binding="{Binding LastAccessTime}" />
+
+                    </DataGrid.Columns>
+                </DataGrid>
+            </TabItem>
+
+        </TabControl>
         </Grid>
     </ScrollViewer>
 </ui:Page>
