+++ src/PresenceLight.Core/Lights/CustomApiService/SetColor/SetColorCommand.cs
@@ -6,6 +6,6 @@ namespace PresenceLight.Core.CustomApiServices
     public class SetColorCommand : IRequest<string>
     {
         public string Availability { get; set; }
-        public string Activity { get; set; }
+        public string? Activity { get; set; }
     }
 }
