+++ src/PresenceLight.Core/Lights/HueServices/SetColor/SetColorCommand.cs
@@ -7,7 +7,7 @@ namespace PresenceLight.Core.HueServices
     public class SetColorCommand : IRequest<Unit>
     {
         public string Availability { get; set; }
-        public string Activity { get; set; }
+        public string? Activity { get; set; }
         public string LightID { get; set; }
     }
 }
