+++ src/PresenceLight.Core/Lights/LifxServices/SetColor/SetColorCommand.cs
@@ -6,8 +6,8 @@ namespace PresenceLight.Core.LifxServices
 {
     public class SetColorCommand : IRequest<Unit>
     {
-        public string Availability { get; set; }
-        public string Activity { get; set; }
+        public string? Availability { get; set; }
+        public string? Activity { get; set; }
         public string? LightId { get; set; }
         public string ApiKey { get; set; }
     }
