+++ src/PresenceLight.Core/Lights/YeelightServices/SetColor/SetColorCommand.cs
@@ -6,8 +6,8 @@ namespace PresenceLight.Core.YeelightServices
 {
     public class SetColorCommand : IRequest<Unit>
     {
-        public string Availability { get; set; }
-        public string Activity { get; set; }
-        public string LightId { get; set; }
+        public string? Availability { get; set; }
+        public string? Activity { get; set; }
+        public string? LightId { get; set; }
     }
 }
